+++ tests/test_agent.py
@@ -10,8 +10,8 @@ from langchain.chat_models import ChatOpenAI
 from langchain.llms import FakeListLLM
 from langchain.memory.chat_memory import BaseChatMemory
 
-from chat_end_detector import SmartMemoryCleaner
-from memory import SavableSummaryBufferMemoryWithDates
+from agents.chat_end_detector import SmartMemoryCleaner
+from agents.memory import SavableSummaryBufferMemoryWithDates
 
 
 def add_test_messages(memory: BaseChatMemory):
